Initial build
The application is built with a combination of react, redux, bootstrap, and sass. I have a store set with a list of quotes as well as dynamic color. The two reducer functions work to grab a random key from the list and save them in the state. The styling is performed with mostly bootstrap with some custom variables. The react is only returning one base function App(), that returns a single container based application.