Updated styles
Added font styling, as well as new variables for bootstrap. Fixed issue with preloading the next color in line for the background, as well as some other minor stylistic changes.  Added the base X button, and will add functionality to create a post on X  with the current quote as the body of the tweet in my next update.

Lastly, i removed some of the base junk associated with the base React app.